be architecture agnostic

+++ pkg/lvm/nodeserver.go
@@ -264,8 +264,8 @@ func (ns *nodeServer) NodeGetVolumeStats(ctx context.Context, in *csi.NodeGetVol
 		return nil, err
 	}
 
-	diskFree := int64(fs.Bfree) * fs.Bsize   // nolint:gosec
-	diskTotal := int64(fs.Blocks) * fs.Bsize // nolint:gosec
+	diskFree := int64(fs.Bfree) * int64(fs.Bsize)   // nolint:gosec
+	diskTotal := int64(fs.Blocks) * int64(fs.Bsize) // nolint:gosec
 
 	inodesFree := int64(fs.Ffree)  // nolint:gosec
 	inodesTotal := int64(fs.Files) // nolint:gosec
